lxvitals-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psutil==5.9.6
PySensors==0.0.4
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
pydantic-settings===2.8.1
ping3==4.0.8
speedtest-cli==2.1.3